+++ internal/controllers/configuration/zz_generated_controllers.go
@@ -613,6 +613,91 @@ func (r *KongV1KongIngressReconciler) Reconcile(ctx context.Context, req ctrl.Re
 }
 
 // -----------------------------------------------------------------------------
+// KongV1beta1 KongUpstreamPolicy - Reconciler
+// -----------------------------------------------------------------------------
+
+// KongV1beta1KongUpstreamPolicyReconciler reconciles KongUpstreamPolicy resources
+type KongV1beta1KongUpstreamPolicyReconciler struct {
+	client.Client
+
+	Log              logr.Logger
+	Scheme           *runtime.Scheme
+	DataplaneClient  controllers.DataPlane
+	CacheSyncTimeout time.Duration
+}
+
+var _ controllers.Reconciler = &KongV1beta1KongUpstreamPolicyReconciler{}
+
+// SetupWithManager sets up the controller with the Manager.
+func (r *KongV1beta1KongUpstreamPolicyReconciler) SetupWithManager(mgr ctrl.Manager) error {
+	c, err := controller.New("KongV1beta1KongUpstreamPolicy", mgr, controller.Options{
+		Reconciler: r,
+		LogConstructor: func(_ *reconcile.Request) logr.Logger {
+			return r.Log
+		},
+		CacheSyncTimeout: r.CacheSyncTimeout,
+	})
+	if err != nil {
+		return err
+	}
+	return c.Watch(
+		source.Kind(mgr.GetCache(), &kongv1beta1.KongUpstreamPolicy{}),
+		&handler.EnqueueRequestForObject{},
+	)
+}
+
+// SetLogger sets the logger.
+func (r *KongV1beta1KongUpstreamPolicyReconciler) SetLogger(l logr.Logger) {
+	r.Log = l
+}
+
+//+kubebuilder:rbac:groups=configuration.konghq.com,resources=kongupstreampolicies,verbs=get;list;watch
+//+kubebuilder:rbac:groups=configuration.konghq.com,resources=kongupstreampolicies/status,verbs=get;update;patch
+
+// Reconcile processes the watched objects
+func (r *KongV1beta1KongUpstreamPolicyReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+	log := r.Log.WithValues("KongV1beta1KongUpstreamPolicy", req.NamespacedName)
+
+	// get the relevant object
+	obj := new(kongv1beta1.KongUpstreamPolicy)
+
+	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
+		if apierrors.IsNotFound(err) {
+			obj.Namespace = req.Namespace
+			obj.Name = req.Name
+
+			return ctrl.Result{}, r.DataplaneClient.DeleteObject(obj)
+		}
+		return ctrl.Result{}, err
+	}
+	log.V(util.DebugLevel).Info("reconciling resource", "namespace", req.Namespace, "name", req.Name)
+
+	// clean the object up if it's being deleted
+	if !obj.DeletionTimestamp.IsZero() && time.Now().After(obj.DeletionTimestamp.Time) {
+		log.V(util.DebugLevel).Info("resource is being deleted, its configuration will be removed", "type", "KongUpstreamPolicy", "namespace", req.Namespace, "name", req.Name)
+
+		objectExistsInCache, err := r.DataplaneClient.ObjectExists(obj)
+		if err != nil {
+			return ctrl.Result{}, err
+		}
+		if objectExistsInCache {
+			if err := r.DataplaneClient.DeleteObject(obj); err != nil {
+				return ctrl.Result{}, err
+			}
+			return ctrl.Result{Requeue: true}, nil // wait until the object is no longer present in the cache
+		}
+		return ctrl.Result{}, nil
+	}
+
+	// update the kong Admin API with the changes
+	if err := r.DataplaneClient.UpdateObject(obj); err != nil {
+		return ctrl.Result{}, err
+	}
+
+	return ctrl.Result{}, nil
+}
+
+// -----------------------------------------------------------------------------
 // KongV1 KongPlugin - Reconciler
 // -----------------------------------------------------------------------------
